+++ perf/tests/test_text_runner.py
@@ -87,7 +87,6 @@ class TestTextRunner(unittest.TestCase):
 
         self.assertEqual(runner.args.loops, 10 ** 5)
         self.assertEqual(result.loops, 10 ** 5)
-        self.assertEqual(result.get_metadata()['loops'], '10^5')
 
         self.assertIn('calibration: 1 loop: 1.00 us\n'
                       'calibration: 10 loops: 10.00 us\n'
@@ -113,7 +112,6 @@ class TestTextRunner(unittest.TestCase):
 
         self.assertEqual(runner.args.loops, 10 ** 3)
         self.assertEqual(result.loops, 10 ** 3)
-        self.assertEqual(result.get_metadata()['loops'], '1000')
 
     def test_json_file_raw(self):
         with tempfile.NamedTemporaryFile('wb+') as tmp: